bharatwitness/requirements.txt
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
pyserini>=0.21.0
paddleocr>=2.7.0
pillow>=9.0.0
opencv-python>=4.7.0
pypdf>=3.0.0
rapidfuzz>=3.0.0
uvicorn>=0.22.0
fastapi>=0.100.0
orjson>=3.9.0
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
rich>=13.0.0
tqdm>=4.65.0
python-levenshtein>=0.21.0
pyyaml>=6.0
ujson>=5.7.0
rank-bm25>=0.2.2
regex>=2023.0.0
dateparser>=1.1.8
python-iso639>=2022.7.25
nltk>=3.8.0
pytest>=7.4.0
layoutparser>=0.3.4
scikit-learn>=1.3.0
pymupdf>=1.23.0
